requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.12
groq==0.11.0
python-telegram-bot==21.6
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2